Move WgpuWrapper back to where it was
